+++ Bugsnag/Client/BugsnagClient.m
@@ -28,35 +28,38 @@
 
 #import "BugsnagClient.h"
 
+#import "BSGConnectivity.h"
+#import "BSGJSONSerialization.h"
+#import "BSGSerialization.h"
+#import "BSG_KSCrash.h"
+#import "BSG_KSCrashC.h"
+#import "BSG_KSCrashReport.h"
+#import "BSG_KSCrashState.h"
+#import "BSG_KSCrashType.h"
+#import "BSG_KSMach.h"
+#import "BSG_KSSystemInfo.h"
+#import "BSG_RFC3339DateTool.h"
+#import "Bugsnag.h"
+#import "Bugsnag.h"
 #import "BugsnagBreadcrumbs.h"
 #import "BugsnagClientInternal.h"
-#import "BSGConnectivity.h"
-#import "Bugsnag.h"
-#import "Private.h"
+#import "BugsnagCollections.h"
 #import "BugsnagCrashSentry.h"
+#import "BugsnagError+Private.h"
+#import "BugsnagErrorTypes.h"
+#import "BugsnagEvent+Private.h"
 #import "BugsnagHandledState.h"
-#import "BugsnagLogger.h"
 #import "BugsnagKeys.h"
+#import "BugsnagLogger.h"
+#import "BugsnagMetadataInternal.h"
+#import "BugsnagNotifier.h"
+#import "BugsnagPluginClient.h"
 #import "BugsnagSessionTracker.h"
 #import "BugsnagSessionTrackingApiClient.h"
-#import "BugsnagPluginClient.h"
-#import "BugsnagSystemState.h"
-#import "BSG_RFC3339DateTool.h"
-#import "BSG_KSCrashC.h"
-#import "BSG_KSCrashType.h"
-#import "BSG_KSCrashState.h"
-#import "BSG_KSSystemInfo.h"
-#import "BSG_KSMach.h"
-#import "BSGSerialization.h"
-#import "Bugsnag.h"
-#import "BugsnagErrorTypes.h"
-#import "BugsnagNotifier.h"
-#import "BugsnagMetadataInternal.h"
 #import "BugsnagStateEvent.h"
-#import "BugsnagCollections.h"
-#import "BSG_KSCrashReport.h"
-#import "BSG_KSCrash.h"
-#import "BSGJSONSerialization.h"
+#import "BugsnagSystemState.h"
+#import "BugsnagThread+Private.h"
+#import "Private.h"
 
 #if BSG_PLATFORM_IOS || BSG_PLATFORM_TVOS
 #define BSGOOMAvailable 1
@@ -119,10 +122,6 @@ NSDictionary *BSGParseDeviceMetadata(NSDictionary *event);
 @interface BugsnagSession ()
 @property NSUInteger unhandledCount;
 @property NSUInteger handledCount;
-@end
-
-@interface BugsnagThread ()
-+ (NSMutableArray *)serializeThreads:(NSArray<BugsnagThread *> *)threads;
 @end
 
 @interface BugsnagAppWithState ()
@@ -326,31 +325,6 @@ void BSGWriteSessionCrashData(BugsnagSession *session) {
 - (BOOL)shouldRecordBreadcrumbType:(BSGBreadcrumbType)type;
 @end
 
-@interface BugsnagEvent ()
-@property(readonly, copy, nonnull) NSDictionary *overrides;
-@property(readwrite) NSUInteger depth;
-@property(readonly, nonnull) BugsnagHandledState *handledState;
-@property (nonatomic, strong) BugsnagMetadata *metadata;
-- (void)setOverrideProperty:(NSString *)key value:(id)value;
-- (NSDictionary *)toJson;
-- (instancetype)initWithApp:(BugsnagAppWithState *)app
-                     device:(BugsnagDeviceWithState *)device
-               handledState:(BugsnagHandledState *)handledState
-                       user:(BugsnagUser *)user
-                   metadata:(BugsnagMetadata *)metadata
-                breadcrumbs:(NSArray<BugsnagBreadcrumb *> *)breadcrumbs
-                     errors:(NSArray<BugsnagError *> *)errors
-                    threads:(NSArray<BugsnagThread *> *)threads
-                    session:(BugsnagSession *)session;
-@end
-
-@interface BugsnagError ()
-- (instancetype)initWithErrorClass:(NSString *)errorClass
-                      errorMessage:(NSString *)errorMessage
-                         errorType:(BSGErrorType)errorType
-                        stacktrace:(NSArray<BugsnagStackframe *> *)stacktrace;
-@end
-
 @interface BugsnagSessionTracker ()
 @property(nonatomic) NSString *codeBundleId;
 @end
@@ -394,8 +368,7 @@ NSString *_lastOrientation = nil;
         self.extraRuntimeInfo = [NSMutableDictionary new];
         self.metadataLock = [[NSLock alloc] init];
         self.crashSentry = [BugsnagCrashSentry new];
-        self.errorReportApiClient = [[BugsnagErrorReportApiClient alloc] initWithConfig:configuration
-                                                                              queueName:@"Error API queue"];
+        self.errorReportApiClient = [[BugsnagErrorReportApiClient alloc] initWithSession:configuration.session queueName:@"Error API queue"];
         bsg_g_bugsnag_data.onCrash = (void (*)(const BSG_KSCrashReportWriter *))self.configuration.onCrashHandler;
 
         static dispatch_once_t once_t;
@@ -824,11 +797,9 @@ NSString *const BSGBreadcrumbLoadedMessage = @"Bugsnag loaded";
             [strongSelf flushPendingReports];
         }
 
-        [self addAutoBreadcrumbOfType:BSGBreadcrumbTypeState
-                          withMessage:@"Connectivity changed"
-                          andMetadata:@{
-                              @"type"  :  connectionType
-                          }];
+        [strongSelf addAutoBreadcrumbOfType:BSGBreadcrumbTypeState
+                                withMessage:@"Connectivity changed"
+                                andMetadata:@{@"type": connectionType}];
     }];
 }