+++ Tests/BugsnagThreadSerializationTest.m
@@ -3,17 +3,11 @@
 // Copyright (c) 2018 Bugsnag. All rights reserved.
 //
 
-#import <Foundation/Foundation.h>
 #import <XCTest/XCTest.h>
 
-#import "BugsnagEvent.h"
+#import "BugsnagEvent+Private.h"
 
 @interface BugsnagThreadSerializationTest : XCTestCase
-@end
-
-@interface BugsnagEvent ()
-- (NSDictionary *)toJson;
-- (instancetype)initWithKSReport:(NSDictionary *)report;
 @end
 
 @implementation BugsnagThreadSerializationTest