+++ Tests/BugsnagUserTest.m
@@ -9,20 +9,12 @@
 #import <XCTest/XCTest.h>
 
 #import "BugsnagUser.h"
-#import "BugsnagEvent.h"
-
-@interface BugsnagEvent ()
-- (instancetype)initWithKSReport:(NSDictionary *)report;
-@end
+#import "BugsnagEvent+Private.h"
 
 @interface BugsnagUser ()
 - (instancetype)initWithDictionary:(NSDictionary *)dict;
 - (instancetype)initWithUserId:(NSString *)userId name:(NSString *)name emailAddress:(NSString *)emailAddress;
 - (NSDictionary *)toJson;
-@end
-
-@interface BugsnagEvent ()
-- (instancetype)initWithKSReport:(NSDictionary *)report;
 @end
 
 @interface BugsnagUserTest : XCTestCase