+++ CHANGELOG.md
@@ -8,6 +8,9 @@ Changelog
 * Error and session deliveries that fail with an HTTP error status code will now be retried if appropriate.
   [#895](https://github.com/bugsnag/bugsnag-cocoa/pull/895)
 
+* Fixed a rare crash in BugsnagClient's network connectivity observer block.
+  [#904](https://github.com/bugsnag/bugsnag-cocoa/pull/904)
+
 ## 6.2.5 (2020-11-18)
 
 ### Bug fixes